Remove the useless components , initiate empty varibles with null and make some changes
