+++ src/App.js
@@ -1,16 +1,14 @@
 import React from 'react';
 import NavBar from "./Components/Navbar";
 import {StylesProvider} from "@material-ui/core/styles";
-import ListOfUsers from "./Components/table";
+import UsersList from "./Components/usersList";
 
 function App() {
     return (
-        <div className="App">
-            <StylesProvider injectFirst>
-                <NavBar/>
-                <ListOfUsers/>
-            </StylesProvider>
-        </div>
+        <StylesProvider injectFirst>
+            <NavBar/>
+            <UsersList/>
+        </StylesProvider>
     );
 }
